+++ src/UnitTestEx/Abstractions/TestFrameworkImplementor.cs
@@ -2,6 +2,7 @@
 
 using Microsoft.Extensions.Logging;
 using System;
+using System.Diagnostics.CodeAnalysis;
 
 namespace UnitTestEx.Abstractions
 {