+++ src/UnitTestEx/Functions/FunctionTesterBase.cs
@@ -233,10 +233,9 @@ namespace UnitTestEx.Functions
         /// <param name="httpMethod">The <see cref="HttpMethod"/>.</param>
         /// <param name="requestUri">The requuest uri.</param>
         /// <param name="requestOptions">The optional <see cref="Ceh.HttpRequestOptions"/>.</param>
-        /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <returns>The <see cref="HttpRequest"/>.</returns>
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri = null, Ceh.HttpRequestOptions? requestOptions = null, params IHttpArg[] args)
-            => CreateHttpRequest(httpMethod, requestUri, null, requestOptions, args);
+        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri = null, Ceh.HttpRequestOptions? requestOptions = null)
+            => CreateHttpRequest(httpMethod, requestUri, null, requestOptions);
 
         /// <summary>
         /// Creates a new <see cref="HttpRequest"/> with <i>optional</i> <paramref name="body"/> as <see cref="HttpRequest.ContentType"/> of <see cref="MediaTypeNames.Text.Plain"/>.
@@ -255,9 +254,8 @@ namespace UnitTestEx.Functions
         /// <param name="requestUri">The requuest uri.</param>
         /// <param name="body">The optional body content.</param>
         /// <param name="requestOptions">The optional <see cref="Ceh.HttpRequestOptions"/>.</param>
-        /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <returns>The <see cref="HttpRequest"/>.</returns>
-        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri = null, string? body = null, Ceh.HttpRequestOptions? requestOptions = null, params IHttpArg[] args)
+        public HttpRequest CreateHttpRequest(HttpMethod httpMethod, string? requestUri = null, string? body = null, Ceh.HttpRequestOptions? requestOptions = null)
         {
             if (httpMethod == HttpMethod.Get && body != null)
                 Implementor.CreateLogger("FunctionTesterBase").LogWarning("A payload within a GET request message has no defined semantics; sending a payload body on a GET request might cause some existing implementations to reject the request (see https://www.rfc-editor.org/rfc/rfc7231).");
@@ -273,18 +271,13 @@ namespace UnitTestEx.Functions
             context.Request.Host = new HostString(uri.Host);
             context.Request.Path = uri.LocalPath;
 
-            // Extend the query string from the IHttpArgs.
-            var qs = new QueryString(uri.Query);
-            foreach (var arg in (args ??= Array.Empty<IHttpArg>()).Where(x => x != null))
-            {
-                qs = arg.AddToQueryString(qs, JsonSerializer);
-            }
-
             // Extend the query string to include additional options.
+            var qs = new QueryString(uri.Query);
             if (requestOptions != null)
                 qs = requestOptions.AddToQueryString(qs);
 
             context.Request.QueryString = qs;
+            context.Request.ApplyETag(requestOptions?.ETag);
 
             if (body != null)
             {
@@ -312,14 +305,13 @@ namespace UnitTestEx.Functions
         /// <param name="requestUri">The requuest uri.</param>
         /// <param name="value">The value to JSON serialize.</param>
         /// <param name="requestOptions">The optional <see cref="Ceh.HttpRequestOptions"/>.</param>
-        /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <returns>The <see cref="HttpRequest"/>.</returns>
-        public HttpRequest CreateJsonHttpRequest(HttpMethod httpMethod, string? requestUri, object value, Ceh.HttpRequestOptions? requestOptions = null, params IHttpArg[] args)
+        public HttpRequest CreateJsonHttpRequest(HttpMethod httpMethod, string? requestUri, object value, Ceh.HttpRequestOptions? requestOptions = null)
         {
             if (httpMethod == HttpMethod.Get)
                 Implementor.CreateLogger("FunctionTesterBase").LogWarning("A payload within a GET request message has no defined semantics; sending a payload body on a GET request might cause some existing implementations to reject the request (see https://www.rfc-editor.org/rfc/rfc7231).");
 
-            var hr = CreateHttpRequest(httpMethod, requestUri, requestOptions, args);
+            var hr = CreateHttpRequest(httpMethod, requestUri, requestOptions);
             hr.Body = new MemoryStream(JsonSerializer.SerializeToBinaryData(value).ToArray());
             hr.ContentType = MediaTypeNames.Application.Json;
             return hr;
@@ -344,10 +336,9 @@ namespace UnitTestEx.Functions
         /// <param name="requestUri">The requuest uri.</param>
         /// <param name="resourceName">The embedded resource name (matches to the end of the fully qualifed resource name).</param>
         /// <param name="requestOptions">The optional <see cref="Ceh.HttpRequestOptions"/>.</param>
-        /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <returns>The <see cref="HttpRequest"/>.</returns>
-        public HttpRequest CreateJsonHttpRequestFromResource<TAssembly>(HttpMethod httpMethod, string? requestUri, string resourceName, Ceh.HttpRequestOptions? requestOptions = null, params IHttpArg[] args)
-            => CreateJsonHttpRequestFromResource(httpMethod, requestUri, resourceName, typeof(TAssembly).Assembly, requestOptions, args);
+        public HttpRequest CreateJsonHttpRequestFromResource<TAssembly>(HttpMethod httpMethod, string? requestUri, string resourceName, Ceh.HttpRequestOptions? requestOptions = null)
+            => CreateJsonHttpRequestFromResource(httpMethod, requestUri, resourceName, typeof(TAssembly).Assembly, requestOptions);
 
         /// <summary>
         /// Creates a new <see cref="HttpRequest"/> using the JSON formatted embedded resource as the content (<see cref="MediaTypeNames.Application.Json"/>).
@@ -368,14 +359,13 @@ namespace UnitTestEx.Functions
         /// <param name="resourceName">The embedded resource name (matches to the end of the fully qualifed resource name).</param>
         /// <param name="assembly">The <see cref="Assembly"/> that contains the embedded resource; defaults to <see cref="Assembly.GetEntryAssembly()"/>.</param>
         /// <param name="requestOptions">The optional <see cref="Ceh.HttpRequestOptions"/>.</param>
-        /// <param name="args">Zero or more <see cref="IHttpArg"/> objects for <paramref name="requestUri"/> templating, query string additions, and content body specification.</param>
         /// <returns>The <see cref="HttpRequest"/>.</returns>
-        public HttpRequest CreateJsonHttpRequestFromResource(HttpMethod httpMethod, string? requestUri, string resourceName, Assembly assembly, Ceh.HttpRequestOptions? requestOptions = null, params IHttpArg[] args)
+        public HttpRequest CreateJsonHttpRequestFromResource(HttpMethod httpMethod, string? requestUri, string resourceName, Assembly assembly, Ceh.HttpRequestOptions? requestOptions = null)
         {
             if (httpMethod == HttpMethod.Get)
                 Implementor.CreateLogger("FunctionTesterBase").LogWarning("A payload within a GET request message has no defined semantics; sending a payload body on a GET request might cause some existing implementations to reject the request (see https://www.rfc-editor.org/rfc/rfc7231).");
 
-            var hr = CreateHttpRequest(httpMethod, requestUri, requestOptions, args);
+            var hr = CreateHttpRequest(httpMethod, requestUri, requestOptions);
             hr.Body = new MemoryStream(Encoding.UTF8.GetBytes(Resource.GetJson(resourceName, assembly ?? Assembly.GetCallingAssembly())));
             hr.ContentType = MediaTypeNames.Application.Json;
             return hr;